+++ CodeSolution/Code/AEngine_Graphics/Models/Animation/AnimationPlayer.cpp
@@ -441,7 +441,7 @@ AEResult AnimationPlayer::Update(const TimerParams& timerParams, const glm::mat4
 
     AEResult ret = AEResult::Ok;
 
-    float elapsedTime = (float)timerParams.m_ElapsedTime;
+    float elapsedTime = timerParams.m_ElapsedTime;
 
     m_CurrentAnimation->Update(timerParams);
